Use LESS; Add base styles; Streamline workflow
