+++ head.html
@@ -1,5 +1,5 @@
 <!-- THIS IS THE DEFAULT FILE CONTENT -->
-<link rel="stylesheet" href="/styles.css"/>
+<link rel="stylesheet" href="/dist/app.min.css"/>
 <link id="favicon" rel="icon" type="image/svg+xml" href="/icons/adobe.svg">
 <link rel="stylesheet" href="/hlx_fonts/pnv6nym.css"/>
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
