+++ package.json
@@ -3,8 +3,12 @@
   "version": "1.0.0",
   "description": "Adobe Federated Publishing",
   "scripts": {
-    "lint": "./node_modules/.bin/eslint .",
-    "buildJS": "rollup --config"
+    "build:js": "./node_modules/.bin/eslint . && rollup --config",
+    "watch:js": "onchange './scripts/**/*.js' -- npm run build:js",
+    "build:css": "lessc ./styles/main.less ./dist/app.css && minify ./dist/app.css > ./dist/app.min.css",
+    "watch:css": "onchange './styles/**/*.less' -- npm run build:css",
+    "app:build": "npm run build:js && npm run build:css",
+    "app:watch": "concurrently 'npm:watch:js' 'npm:watch:css'"
   },
   "repository": {
     "type": "git",
@@ -18,10 +22,18 @@
   "homepage": "https://github.com/adobe/fedpub#readme",
   "devDependencies": {
     "@adobe/eslint-config-helix": "1.1.3",
+    "concurrently": "^5.3.0",
     "eslint": "7.15.0",
+    "eslint-plugin-file-progress": "^1.1.1",
     "eslint-plugin-header": "3.1.0",
     "eslint-plugin-import": "2.22.1",
     "rollup": "2.37.0",
-    "rollup-plugin-terser": "7.0.2"
+    "rollup-plugin-terser": "7.0.2",
+    "less": "^4.1.0",
+    "minify": "^6.0.1",
+    "onchange": "^7.1.0"
+  },
+  "dependencies": {
+    "reset-css": "^5.0.1"
   }
 }
