+++ rollup.config.js
@@ -1,7 +1,7 @@
 import { terser } from 'rollup-plugin-terser';
 
 export default {
-    input: './scripts.js',
+    input: './scripts/main.js',
     output: {
         file: './dist/app.min.js',
         format: 'iife',
