+++ scripts/modules/lang/isNonEmptyString.js
