+++ scripts/modules/lang/stringToClassname.js
