--- styles.css
@@ -1,62 +1,0 @@
-body{
-    background-color: lightgrey;
-} 
-
-img {
-    max-width: 100%;
-}
-
-* {
-    box-sizing: border-box;
-}
-
-main .section-wrapper > div {
-    margin: auto;
-    max-width: 780px;
-}
-
-main a:any-link {
-	color: #1473e6;
-	text-decoration: none;
-}
-
-main .button {
-	border-width: 2px;
-	border-style: solid;
-	border-radius: 16px;
-	min-height: 32px;
-	height: auto;
-	min-width: 72px;
-	padding: 4px 14px;
-	padding-bottom: 4.5px;
-	padding-top: 3.5px;
-	font-size: 14px;
-	font-weight: 700;
-	transition: background .13s ease-out, border-color .13s ease-out, color .13s ease-out, box-shadow .13s ease-out;
-  line-height: 1.3;
-  -webkit-appearance: none;
-}
-
-main .button.primary {
-	background-color: #1473e6;
-	border-color: #1473e6;
-	color: #fff;
-}
-
-main .button.primary:hover {
-	background-color: #0d66d0;
-	border-color: #0d66d0;
-	color: #fff;
-}
-
-main .button.secondary {
-	background-color: transparent;
-	border-color: #505050;
-	color: #505050;
-}
-
-main .button.secondary:hover {
-	background-color: #505050;
-	border-color: #505050;
-	color: #fff;
-}
