--- styles/blocks/cards.css
@@ -1,14 +1,0 @@
-main div.cards {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-} 
-
-main div.cards > div{
-    
-    display: flex;
-    flex-direction: column;
-    width: 250px;
-    padding: 32px;
-}
