+++ styles/variables.less
@@ -1,0 +1,27 @@
+// Breakpoints
+@breakpoint-mobile: 800px;
+
+// Colours
+@background-body: #fff;
+@color-link-default: #1473e6;
+
+// Typography
+@fontFamily-default: adobe-clean, sans-serif;
+@lineHeight-default: 1.45;
+@fontWeight-headlines: 700;
+
+// Dimensions
+@gutter-content: 50px;
+@width-content-outer: 800px;
+@width-content-inner: 600px;
+
+// Callout
+@spacing-callout-default: 64px;
+@gutter-callout-mobile: 20px;
+@gutter-callout-desktop: 100px;
+@size-callout-branding: 8px;
+@background-callout: #f5f5f5;
+@color-callout-cc--from: #ffa701;
+@color-callout-cc--to: #e329e0;
+@gradient-callout-cc--mobile: linear-gradient(90deg, @color-callout-cc--from, @color-callout-cc--to);
+@gradient-callout-cc--desktop: linear-gradient(@color-callout-cc--from, @color-callout-cc--to);
